refactor: 패키지 구조 개선
- domain 영역의 port를 사용하거나 구현하는 경우 adapter 영역으로 설정
- adapter 영역에서만 사용하는 경우 infrastructure 영역으로 설정
